requirements.txt
SpeechRecognition
pyttsx3
pyaudio
python-dotenv
pvporcupine
sounddevice
openai
pyautogui
llama-cpp-python
